0602 - 관리자: 파일 구조 수정
