+++ src/main/java/com/triplan/planner/admin/service/CommunityService.java
@@ -1,57 +1,17 @@
 package com.triplan.planner.admin.service;
 
 import com.triplan.planner.admin.domain.Community;
-import com.triplan.planner.admin.repository.CommunityMapper;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Map;
 
-@Service
-public class CommunityService {
+public interface CommunityService {
+    List<Community> getPostsByCategory(String category);
 
-    @Autowired
-    private CommunityMapper communityMapper;
+    List<Community> getPagedPostsByCategory(String category, String searchType, String searchQuery, int startRow, int size);
 
-    public List<Community> getPostsByCategory(String category) {
-        return communityMapper.getPostsByCategory(category);
-    }
+    int getCountByCategory(String category, String searchType, String searchQuery);
 
-    public List<Community> getPagedPostsByCategory(String category, String searchType, String searchQuery, int startRow, int size) {
-        Map<String, Object> params = Map.of(
-                "category", category,
-                "searchType", searchType,
-                "searchQuery", searchQuery,
-                "startRow", startRow,
-                "size", size
-        );
-        return communityMapper.getPagedPostsByCategory(params);
-    }
+    void deletePosts(List<Integer> boardIdxArray, String category);
 
-    public int getCountByCategory(String category, String searchType, String searchQuery) {
-        Map<String, Object> params = Map.of(
-                "category", category,
-                "searchType", searchType,
-                "searchQuery", searchQuery
-        );
-        return communityMapper.getCountByCategory(params);
-    }
-
-    public void deletePosts(List<Integer> boardIdxArray, String category) {
-        Map<String, Object> params = Map.of(
-                "boardIdxArray", boardIdxArray,
-                "category", category
-        );
-        communityMapper.deletePosts(params);
-    }
-
-    public List<Community> searchPosts(String category, String searchType, String searchQuery) {
-        Map<String, Object> params = Map.of(
-                "category", category,
-                "searchType", searchType,
-                "searchQuery", searchQuery
-        );
-        return communityMapper.searchPosts(params);
-    }
+    List<Community> searchPosts(String category, String searchType, String searchQuery);
 }
